Reworked PhotoCamera
Reworked PhotoCameraEquipment to use animation state machine and anim notifies to control raising and lowering the camera to fix bug where camera was getting stuck in unusable state when spamming RMB during transitions. Handled situation in which Photo render target is inexplicably invalid to prevent crashing when attempting to display it.