Various Bug Fixes
Added 4k resolution option. Fixed possible crashes associated with null pointers in photo subject AI controllers and behavior trees. Fixed issues with cursor appearing at the edges of the screen and clicking outside the game window when it should be locked to the viewport. Fixed issue with installer pointing to the wrong exe name for start menu shortcut.